requirements.txt
construct
httpx
PyNaCl